feat: navigator 추가 및 분기처리
