+++ src/screens/index.tsx
@@ -1,5 +1,6 @@
-export { default as HomeScreen } from './HomeScreen'
 export { default as LoginScreen } from './LoginScreen'
-export { default as StreakTrackerScreen } from './StreakTrackerScreen'
+export { default as HomeScreen } from './HomeScreen'
 export { default as ProfileScreen } from './ProfileScreen'
+export { default as StreakTrackerScreen } from './StreakTrackerScreen'
+export { default as NavigateScreen } from './NavigateScreen'
 export { default as TestScreen } from './TestScreen'
